// src/api/model/body3.ts
/**
 * LinkIOT Dashboard
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Body3 { 
    /**
     * 设备ID
     */
    deviceId: string;
    /**
     * 新的设备名称
     */
    name?: string;
    /**
     * 新的设备描述
     */
    description?: string;
    /**
     * 新的设备脚本
     */
    script?: string;
}